feat: correct controller.ts
