+++ news-JS/src/components/controller/controller.ts
@@ -1,0 +1,41 @@
+import AppLoader from './appLoader';
+import { controllerFunc, SourcesItem, NewsItem } from '../../types/index'
+
+
+class AppController extends AppLoader {
+  getSources(callback: controllerFunc<SourcesItem>) {
+    super.getResp(
+      {
+        endpoint: 'sources',
+      },
+      callback
+    );
+  }
+
+  getNews(e: Event, callback: controllerFunc<NewsItem>) {
+    let target = e.target as HTMLElement;
+    const newsContainer = e.currentTarget as HTMLElement;
+
+    while (target !== newsContainer) {
+      if (target.classList.contains('source__item')) {
+        const sourceId = target.getAttribute('data-source-id');
+        if (newsContainer.getAttribute('data-source') !== sourceId && sourceId) {
+          newsContainer.setAttribute('data-source', sourceId);
+          super.getResp(
+            {
+              endpoint: 'everything',
+              options: {
+                sources: sourceId,
+              },
+            },
+            callback
+          );
+        }
+        return;
+      }
+      target = target.parentNode as HTMLElement;
+    }
+  }
+}
+
+export default AppController;
